the changes in plt figure that requires integer positions of the figure from numbers import Integral,throat_list=[0], isinstance(np.ceil(np.sqrt(len(throat_list))),Integral) gives False, which makes problem for deprecation refer to https://github.com/matplotlib/matplotlib/commit/4a8aa8ce227e4b823305e2adb616facd1435764b#diff-c3ce78ddcc28c54e04ade080c149c1b1L1267

+++ openpnm/materials/VoronoiFibers.py
@@ -890,7 +890,7 @@ class DelaunayGeometry(GenericGeometry):
             coms = self["throat.centroid"][throat_list]
             incentre = self["throat.incenter"][throat_list]
             inradius = 0.5 * self["throat.indiameter"][throat_list]
-            row_col = np.ceil(np.sqrt(len(throat_list)))
+            row_col = int(np.ceil(np.sqrt(len(throat_list))))
             for i in range(len(throat_list)):
                 if fig is None:
                     fig = plt.figure()
